refactoring: updated the file sturcture
