+++ go.mod
@@ -2,4 +2,4 @@ module todoapp
 
 go 1.22.5
 
-require github.com/google/uuid v1.6.0 // indirect
+require github.com/google/uuid v1.6.0
