+++ models/todo.go
@@ -1,0 +1,9 @@
+package models
+
+type ToDo struct {
+	Id string `json:"id"`
+	Title string `json:"title"`
+	Completed *bool `json:"completed"`
+}
+
+
