+++ server/handlers_test.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"bytes"
@@ -9,21 +9,6 @@ import (
 	"strings"
 	"testing"
 )
-
-func TestMain(t *testing.T) {
-	go main()
-	resp, err := http.Get("http://localhost:8080/todos")
-	if err != nil {
-		t.Errorf("Server failed to start %v", err)
-	}
-	if resp.StatusCode != http.StatusOK {
-		t.Errorf("Sever failed to start %v", resp.Status)
-	}
-
-	if err := http.ListenAndServe(":8080", nil); err == nil {
-		t.Errorf("Server started with port already in use")
-	}
-}
 
 func TestGetAndPostHandler(t *testing.T) {
 	type request struct {
