+++ server/routes.go
@@ -1,0 +1,10 @@
+package server
+
+import (
+	"net/http"
+)
+
+func RegisterRoutes() {
+    http.HandleFunc("/todos", getAndPostHandler)
+    http.HandleFunc("/todos/", putAndDeleteHandler)
+}
